test: Check issues severity
