+++ test/unit/linter.js
@@ -1,115 +1,43 @@
-var knife = require("../../lib/knife");
 const { expect } = require("chai");
-var Linter = require("../../lib/legacy/linter");
+const Linter = require("../../lib/linter");
+const Config = require("../../lib/config");
 
-describe("LegacyLinter", function() {
-  var linter = null;
+describe("Config", function() {
 
-  it("should be a function", function() {
-    expect(Linter).to.be.an.instanceOf(Function);
-  });
-
-  beforeEach(function() {
-    linter = new Linter([
-      {
-        name: "dom",
-        lint: function() {
-          return [];
-        }
-      },
-      {
-        name: "line",
-        lint: function() {
-          return [];
-        }
-      },
-      require("../../lib/rules/free-options.js")
-    ]);
-  });
-
-  describe("lint", function() {
-    var ConstRule = require("../fixtures/const_rule");
-
-    var rule = new ConstRule([
-      {
-        msg: "this is a test",
-        index: 4,
-        line: 2,
-        column: 3
-      },
-      {
-        msg: "this is a test",
-        index: 2
+  this.beforeEach(function() {
+    const foo = {
+      name: "foo",
+      lint(node, options, { report }) {
+        return report({
+          code: "CUSTOM",
+          position: [0, 0]
+        });
       }
-    ]);
-
-    it("Should throw an error when given a nonexistent option", function() {
-      expect(() => linter.lint("f\nfff", { nonopt: 7 }, "nodefault"))
-        .to
-        .throw(`Rule "nonopt" does not exist`);
-    });
-
-    it("should return correct line and column numbers", async function() {
-      linter.rules.addRule(rule);
-      const issues = await linter.lint("f\nfff", "nodefault");
-      expect(issues[0].line).to.be.eql(2);
-      expect(issues[0].column).to.be.eql(3);
-    });
-
-    it("should not truncate output if maxerr is false", async function() {
-      linter.rules.addRule(rule);
-      const issues = await linter.lint("f\nfff", { maxerr: false }, "nodefault");
-      expect(issues).to.have.length(2);
-    });
-
-    it("should not return more than the limit fixed by maxerr", async function() {
-      linter.rules.addRule(rule);
-      const issues = await linter.lint("f\nfff", { maxerr: 1 }, "nodefault");
-      expect(issues).to.have.length(1);
-    });
-
-    it("Should throw an error for non-integer config for maxerr", function() {
-      expect(() => linter.lint("", { maxerr: "five" }, "nodefault"))
-        .to
-        .throw('Configuration for rule "maxerr" is invalid: Expected number got string');
-    });
+    };
+    this.config = new Config({foo});
+    this.rule = this.config.getRule("foo");
   });
 
-  describe("resetRules", function() {
-    it("should return an array of issues", function() {
-      var issue = { msg: "hit" };
-
-      linter.rules.addRule({
-        end: function() {
-          return issue;
-        }
-      });
-
-      var output = linter.resetRules();
-
-      expect(output[0]).to.be.eql(issue);
-    });
+  it('Should report an issue with the "error" severity', async function() {
+    const rule_config = {
+      "foo": "error"
+    };
+    this.config.setRuleConfig(this.rule, rule_config);
+    const linter = new Linter();
+    linter.config = this.config;
+    
+    const issues = await linter.lint('<div></div>');
+    expect(issues[0].severity).to.equal('error');
   });
-
-  // TODO: move these out of this file...
-  describe("shred", function() {
-    it("should return an array", function() {
-      var output = knife.shred("");
-      expect(output).to.be.an.instanceOf(Array);
-    });
-
-    it("should return the proper number of lines", function() {
-      const lines = [
-        "Line1",
-        "Line2",
-        "Line3"
-      ].join("\n");
-      const output = knife.shred(lines);
-
-      expect(output.length).to.be.eql(3);
-      expect(output[0].text).to.equal('Line1\n');
-      expect(output[1].text).to.equal('Line2\n');
-      expect(output[2].text).to.equal('Line3');
-    });
+  it('Should report an issue with the "warning" severity', async function() {
+    const rule_config = {
+      "foo": "warning"
+    };
+    this.config.setRuleConfig(this.rule, rule_config);
+    const linter = new Linter();
+    linter.config = this.config;
+    
+    const issues = await linter.lint('<div></div>');
+    expect(issues[0].severity).to.equal('warning');
   });
 });
